version 1.2
instead of you going back to v1.0 i will list everything here:

views needs fixing, i need it far out but not to much far out, whilst
its still HD as i tried and it just stretched everything.

ive added a new item, i also have an idea for it too, but for now its a
prop, Its a candy cane.

ive done the main menu buttons, some things need changing though, such
as the scale of the buttons, and the click boundary(change whatever you
think is nessicary, you are the coder after all, dont put the buttons to
close together as i need to add more buttons, but i will wait till you
have fixed the size so i can duplicate them lol, im idle).

ive added the throw animation, as explained on facebook i need you to
code it so, when the player presses and holds mouse button, the power
bar will come up, playing the animation up to the frame where his hand
is near his head, then on release when they have it at whatever power
they want, it will throw the snowball, playing the animation in reverse.
if you can, delete the fact that they can change the direction of the
snowballs and make it so they are locked position at either left or
right, as im not to sure how id set up an animation, that allows them to
move the arms freely.

that is all, have fun haha.