v0.0.1
I fixed the snowballs being able to be thrown anywhere. They can now only be thrown straight ahead. The "force" bar for throwing a snowball is a bit complicated and might take me awhile to figure out. I'll add it in when I'm read (won't be for a while). I fixed the resolution. In-game, it's 4k. The menu is only 720p. That's a huge change so some computers might lag a lot on it. I fixed the button scaling and mouse hover boundaries. I also just made the candy cane a bit more performance-bound. You now have to define the direction it's facing in the creation code in the room editor. Change it back if you want, idc. I also just changed a few file names. There ya go.