v0.0.4
this will be your biggest update so far

Added a breakable floor, when destroyed the normal floor on top should
drop to the bottom, giving them access to the switch, try and get this
done if possible, by breakable i mean, ive made a sprite for the floor,
but you will need to make it breakable.
Added first boss (doesnt need anything done just yet)

Added a power up, there will be multiple. this will need coding, i will
need it so when they pick up the powerup their snowballs turn to fire
for a bit(power up usually drops from enemy, but dont need that coding
just yet, just get it working, as for the tutorial it will be there for
them to pick up and test) with this snowball they get the power up for
10 seconds

enemy needs an ai doing, so that is also available to throw snowballs.

problem with ladder that needs fixing, test it and you will see the
problem,

collision needs fixing, as player gets stuck in alot of stuff easy.

health system needs implementing, do it as you please, you dont need the
health bar thats already there, u can change it if you like.

Good luck, there is more needing to be done but those can wait, these
are most important at this moment, please try and get them sorted for
next update, thankyou and have fun!