v0.0.7
I changed the artwork to grayscale.
I added the "Code Puzzle". Press enter when you walk up to the block barrier and type in the code you found.

Here's what we need:
A background for the "Tutorial Complete" room that matches the style of the menu.

We also need a button that says "Menu".

Next we just need to finish up the tutorial.