v0.1.0
Added aberration filter. Open the object for more details.

Finished the tutorial with tool tips. When you complete the tutorial, the game saves that you completed it. So it skips it the next time you click play and goes to a level select screen. We need a background and icons for this. Icons will be "1, 2, 3, etc..." for the level numbers. Each icon will need 2 sub images. An unlocked and a locked sprite.

To access the folder and delete the tutorial save file, go to %appdata% and it should be the first file (the project name). Delete the text file to do the tutorial again.

Add what I said above, and add stuff to the filter if you want. Good luck, have fun!