v0.1.5
Added double jump.
I have no idea what's wrong with the HUD overlapping like that...
I added the first level using a basic button. Don't change the mechanics and names. We will have tooltips saying the name of the level.
Add a background for the level select, and buttons about the same size as the one i put in. We also need a "Level Complete!" background.
Lastly, I added a ladder in level 1 just so we can see the ending.

Let me know if you need anything else.