v0.1.7
What I did / added:

> I fixed Level Select Button.
> Teleporter now goes to Level Complete room.
> Added settings button to main menu (needs proper scaling).
> Added a fullscreen option that toggles fullscreen on or off.
> Adjusted the amount the snowballs do to both enemies and the player.
> Added basic boss AI. The snowballs from the boss follow the player (don't do any damage to the player yet). The only way to destroy them is to hit them with your snowballs.
> Tweaked a few things in the levels.
> Added pause button. It can't take you to the menu yet, but it's very easy to add.
> I could not get the powerup to float. You'll have to research that and try coding it.

TIP FOR AFTER DEMO RELEASE:
-----------------------------------
After we release the demo we need to properly scalr the entire game. Your artwork is very good but is not the same aspect ratio. You have backgrounds and some art that is in 4K while others are super small. We need to keep ALL artwork (even backgrounds) under 2048x2048px. We literally had to change our view to 4K to run the game. This needs to change. Grandmas computer can't even run this game right now. So after we put out the demo we should take some time and scale the game.

Thanks! :)